[feature/select_screen] Theme Merge 충돌 해결 (#26)
