+++ presentation/src/main/java/com/everymeal/presentation/MainActivity.kt
@@ -12,7 +12,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.everymeal.presentation.theme.EveryMeal_AndroidTheme
+import com.everymeal.presentation.ui.theme.EveryMeal_AndroidTheme
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
