--- presentation/src/main/java/com/everymeal/presentation/theme/Color.kt
@@ -1,19 +1,0 @@
-package com.everymeal.presentation.theme
-
-import androidx.compose.ui.graphics.Color
-
-val Purple80 = Color(0xFFD0BCFF)
-val PurpleGrey80 = Color(0xFFCCC2DC)
-val Pink80 = Color(0xFFEFB8C8)
-
-val Purple40 = Color(0xFF6650a4)
-val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
-
-val Gray100 = Color(0xFFF9FAFB)
-val Gray200 = Color(0xFFE5E8EB)
-val Gray300 = Color(0xFFF2F4F6)
-val Gray500 = Color(0xFFB0B8C1)
-var Gray800 = Color(0xFF4E5968)
-
-val Main100 = Color(0xFFFF4848)
