+++ presentation/src/main/java/com/everymeal/presentation/ui/signup/UnivSelectActivity.kt
@@ -6,7 +6,7 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import com.everymeal.presentation.MainActivity
-import com.everymeal.presentation.theme.EveryMeal_AndroidTheme
+import com.everymeal.presentation.ui.theme.EveryMeal_AndroidTheme
 import dagger.hilt.android.AndroidEntryPoint
