+++ presentation/src/main/java/com/everymeal/presentation/ui/splash/SplashActivity.kt
@@ -4,7 +4,7 @@ import android.annotation.SuppressLint
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.everymeal.presentation.theme.EveryMeal_AndroidTheme
+import com.everymeal.presentation.ui.theme.EveryMeal_AndroidTheme
 import com.everymeal.presentation.ui.signup.UnivSelectActivity
 import dagger.hilt.android.AndroidEntryPoint
